@@ -15,37 +15,38 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/cdk": "^6.4.1",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/core": "^6.0.0",
-    "@angular/forms": "^6.0.0",
-    "@angular/http": "^6.0.0",
-    "@angular/material": "^6.4.1",
-    "@angular/platform-browser": "^6.0.0",
-    "@angular/platform-browser-dynamic": "^6.0.0",
-    "@angular/router": "^6.1.0",
-    "@angular/upgrade": "^6.0.0",
+    "@angular/animations": "^6.1.1",
+    "@angular/cdk": "^6.4.2",
+    "@angular/common": "^6.1.1",
+    "@angular/compiler": "^6.1.1",
+    "@angular/core": "^6.1.1",
+    "@angular/forms": "^6.1.1",
+    "@angular/http": "^6.1.1",
+    "@angular/material": "^6.4.2",
+    "@angular/platform-browser": "^6.1.1",
+    "@angular/platform-browser-dynamic": "^6.1.1",
+    "@angular/router": "^6.1.1",
+    "@angular/upgrade": "^6.1.1",
     "angular-in-memory-web-api": "^0.6.0",
     "bootstrap": "^4.1.3",
+    "codelyzer": "^4.4.3",
     "core-js": "^2.5.4",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "ng2-file-upload": "^1.3.0",
-    "popper.js": "^1.14.3",
+    "popper.js": "^1.14.4",
     "rxjs": "^6.0.0",
     "zone.js": "^0.8.24"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.0",
-    "@angular/cli": "^6.0.0",
-    "@angular/compiler-cli": "^6.0.0",
-    "@angular/platform-server": "^6.0.0",
+    "@angular/cli": "^6.1.2",
+    "@angular/compiler-cli": "^6.1.1",
+    "@angular/platform-server": "^6.1.1",
     "@types/jasmine": "~2.8.0",
     "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^6.0.45",
+    "@types/node": "^6.0.116",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
@@ -56,7 +57,7 @@
     "karma-phantomjs-launcher": "^1.0.2",
     "lodash": "^4.16.2",
     "phantomjs-prebuilt": "^2.1.7",
-    "protractor": "~5.3.0",
+    "protractor": "^5.4.0",
     "ts-node": "^5.0.1",
     "tslint": "^5.9.1",
     "typescript": "2.7.2"
