@@ -1,10 +1,11 @@
 import {Component, ElementRef, OnInit} from '@angular/core';
 import {FormBuilder} from '@angular/forms';
 import {Validators} from '@angular/forms';
-import {RestaurantService} from "../../services/restaurant.service";
-import {ImageUploadService} from "../../services/image-upload.service";
-import {ActivatedRoute, Router} from "@angular/router";
+import {RestaurantService} from '../../services/restaurant.service';
+import {ImageUploadService} from '../../services/image-upload.service';
+import {ActivatedRoute, Router} from '@angular/router';
 import {FileUploader} from 'ng2-file-upload';
+import {Http} from "@angular/http";
 import {HttpClient} from "@angular/common/http";
 import {environment} from "../../../environments/environment";
 
