@@ -3,15 +3,16 @@ import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http
 import { Response } from '@angular/http';
 import { Observable } from 'rxjs';
 import { catchError } from 'rxjs/operators';
-import { environment } from '../../environments/environment';
+import { environment} from '../../environments/environment';
+import { ApiService } from './api.service';
 
 @Injectable()
 export class UserService {
 
-  readonly rootUrl = environment.api_url;
+  readonly rootUrl = environment.apiUrl;
   userToken: string;
 
-  constructor(private http: HttpClient) {
+  constructor(private http: HttpClient, private apiService: ApiService) {
     const loggedInToken = localStorage.getItem('userToken');
     if (loggedInToken) {
       this.userToken = loggedInToken;
@@ -34,7 +35,8 @@ export class UserService {
     const data = { 'email': email, 'password': password };
     // tslint:disable-next-line:max-line-length
     const reqHeader = new HttpHeaders(
-      { 'Content-Type': 'application/json', 'Accept': 'application/json', 'Access-Control-Allow-Origin': '*', 'Access-Control-Allow-Headers': 'Authorization, Content-Type' });
+      { 'Content-Type': 'application/json', 'Accept': 'application/json', 'Access-Control-Allow-Origin': '*',
+      'Access-Control-Allow-Headers': 'Authorization, Content-Type' });
 
     return this.http.post(this.rootUrl + 'login', data, { headers: reqHeader });
   }
@@ -54,6 +56,10 @@ export class UserService {
     } else {
       return false;
     }
+  }
+
+  userDetails() {
+    return this.apiService.get('details');
   }
 
   logout() {
