@@ -97,11 +97,13 @@ class RestaurantsController extends Controller
             'state' => $item->address->state->name,
             'category' => $item->categories[0]->name,
             'cuisine' => $item->cuisines[0]->name,
-            'image' => $this->getImages($item)
+            'image' => $this->getImages($item),
         ];
 
         return $values;
     }
+
+    
 
 
 }
