@@ -22,26 +22,29 @@ Route::group(['middleware' => 'auth:api'],
     function () {
 
         //Reviews
+
         Route::get('reviews/post', 'ReviewsController@postRating');
-        Route::post('ratings', 'ReviewsController@postRating');
-        Route::post('reviews', 'ReviewsController@postReview');
+        Route::post('reviews/{id}', 'ReviewsController@postReview');
         Route::get('reviews/{id}', 'ReviewsController@show');
+        Route::get('restaurants/{id}/reviews','ReviewsController@index');
 
-        Route::post('reviews/comment/{id}', 'ReviewsController@createReviewComment');
-        Route::post('reviews/like/{id}', 'ReviewsController@createReviewLike');
 
+        Route::post('reviews/comment/{id}', 'CommentsController@createReviewComment');
+        Route::post('like/{id}', 'LikesController@createReviewLike');
+        //Route::post('/reviews/images/{id}', 'ImagesController@createReviewImage');
 
         Route::get('restaurants', 'RestaurantsController@index');
         Route::post('/restaurants/create', 'RestaurantsController@store');
         Route::get('/restaurants/{id}', 'RestaurantsController@show');
-
-
+        //Route::post('/restaurants/images/{id}', 'ImagesController@createRestaurantImage');
 
         Route::get('/categories', 'CategoriesController@index');
         Route::post('/categories/create', 'CategoriesController@store');
 
         //UserAddress
-        Route::post('/user/address', 'UserAddressController@store');
+        Route::get('/useraddress', 'UserAddressController@index');
+        Route::post('/useraddress', 'UserAddressController@store');
+        Route::put('/useraddress/{id}','UserAddressController@update');
         Route::get('/useraddress/{id}', 'UserAddressController@show');
         Route::delete('useraddress/{id}', 'UserAddressController@destroy');
 
@@ -50,6 +53,13 @@ Route::group(['middleware' => 'auth:api'],
 
        // Route::get('image')
 
+
+        Route::get('/details','PassportController@details');
+        Route::post('/details','PassportController@updateDetails');
+
+        //State
+        Route::get('/states','StateController@index');
+        Route::get('/districts','DistrictController@index');
 
 
     });
